new remove.sh

+++ scripts/remove.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-rm -rf /var/www/html/scripts
-rm -rf /var/www/html/appspec.yml
+[ -d /var/www/html/scripts ] && rm -rf /var/www/html/scripts
+[ -f /var/www/html/appspec.yml ] && rm -rf /var/www/html/appspec.yml
